test: get deleted sizes
