+++ src/test/resources/scripts/sizes/create_sizes.sql
@@ -1,0 +1,4 @@
+INSERT into sizes (size_id, deleted, description, name)
+VALUES ('59985182-9b95-40e2-8f8b-f66769fc6847', 1, 'Small', 'S'),
+       ('cc7f119e-93e0-43f0-8fa4-fcbfd7616265', 0, 'Large', 'L'),
+       ('e206a421-8e9c-45bd-ad53-4da324a9b34e', 1, 'Medium', 'M');
